transfer_bundle_string_in_fragment add NewsActivity.kt add activity_news.xml delete NewsRecyclerActivity.kt delete activity_news_recycler.xml implementation of the logic of passing arguments from the activity to the fragment
