+++ app/src/main/AndroidManifest.xml
@@ -1,10 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="ru.yundon.rss">
 
-    <uses-permission
-        android:name="android.permission.INTERNET" />
+    <uses-permission android:name="android.permission.INTERNET" />
 
     <application
         android:allowBackup="true"
@@ -14,17 +12,20 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.RSS">
         <activity
-            android:name="ru.yundon.rss.presantation.ui.secondactivity.NewsRecyclerActivity"
+            android:name=".presantation.ui.activity.NewsActivity"
             android:exported="false" />
         <activity
-            android:name="ru.yundon.rss.presantation.ui.MainActivity"
+            android:name=".presantation.ui.activity.NewsRecyclerActivity"
+            android:exported="false" />
+        <activity
+            android:name=".presantation.ui.activity.MainActivity"
             android:exported="true">
             <intent-filter>
-                <action
-                    android:name="android.intent.action.MAIN" />
-                <category
-                    android:name="android.intent.category.LAUNCHER" />
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
     </application>
+
 </manifest>
