+++ app/src/main/java/ru/yundon/rss/data/repository/RssRepositoryImpl.kt
@@ -5,11 +5,11 @@ import android.util.Log
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import ru.yundon.rss.data.api.dto.RssItemDto
-import ru.yundon.rss.data.api.response.RssApiClient
+import ru.yundon.rss.data.api.network.RssApiClient
 import ru.yundon.rss.data.database.RssDatabase
 import ru.yundon.rss.data.database.RssDbModel
 import ru.yundon.rss.data.mapper.RssMapper
-import ru.yundon.rss.domain.RssEntity
+import ru.yundon.rss.domain.model.RssEntity
 import ru.yundon.rss.domain.RssRepository
 import ru.yundon.rss.utils.TypeOfNews
