+++ app/src/main/java/ru/yundon/rss/domain/usecases/IsFavoritesUseCase.kt
@@ -1,4 +1,7 @@
-package ru.yundon.rss.domain
+package ru.yundon.rss.domain.usecases
+
+import ru.yundon.rss.domain.RssRepository
+import ru.yundon.rss.domain.model.RssEntity
 
 class IsFavoritesUseCase(private val repository: RssRepository) {
