+++ app/src/main/java/ru/yundon/rss/presantation/adapter/RssDiffCallback.kt
@@ -1,7 +1,7 @@
 package ru.yundon.rss.presantation.adapter
 
 import androidx.recyclerview.widget.DiffUtil
-import ru.yundon.rss.domain.RssEntity
+import ru.yundon.rss.domain.model.RssEntity
 
 class RssDiffCallback: DiffUtil.ItemCallback<RssEntity>() {
     override fun areItemsTheSame(oldItem: RssEntity, newItem: RssEntity): Boolean {
