--- app/src/main/java/ru/yundon/rss/presantation/ui/MainActivity.kt
@@ -1,27 +1,0 @@
-package ru.yundon.rss.presantation.ui
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import androidx.navigation.NavController
-import androidx.navigation.fragment.NavHostFragment
-import androidx.navigation.ui.setupWithNavController
-import ru.yundon.rss.R
-import ru.yundon.rss.databinding.ActivityMainBinding
-
-class MainActivity: AppCompatActivity() {
-
-    private lateinit var binding: ActivityMainBinding
-    private lateinit var navController: NavController
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        binding.apply {
-            navController = (supportFragmentManager.findFragmentById(R.id.mainNavHostFragment) as NavHostFragment).navController
-            bottomNavigation.setupWithNavController(navController)
-        }
-    }
-}
-
